mngr: split agent address resolution into "find" and "ensure live"
Single-agent subcommands previously called find_one_agent with two
ill-defined flags (is_start_desired, skip_agent_state_check) that
silently conflated host- and agent-state decisions. The default required
a fully live agent, which caused push, pull, and exec-single to fail
against stopped agents on online hosts -- even though their operations
only need the host.

Restructure resolution into two explicit phases. find_one_agent now
returns (DiscoveredHost, DiscoveredAgent) -- pure metadata. Two new
helpers in cli/agent_utils.py compose that into live interfaces:
ensure_host_started_and_resolve_agent (host live, agent state ignored)
for push, pull, exec-single, provision, and rename;
ensure_host_and_agent_started (host live + agent running) for connect
and capture. Both take a single allow_auto_start flag driven by --start.

User-visible: push, pull, and provision no longer fail against stopped
agents; push, pull, provision, and rename gain a --start/--no-start
flag (default --start). The agent-selector TUI is extracted from
connect.py into a new agent_selector.py to break the import cycle the
new structure exposes.
