+++ libs/mngr/imbue/mngr/cli/capture.py
@@ -6,6 +6,7 @@ from click_option_group import optgroup
 from loguru import logger
 
 from imbue.mngr.cli.address_params import AGENT_ADDRESS
+from imbue.mngr.cli.agent_utils import ensure_host_and_agent_started
 from imbue.mngr.cli.agent_utils import find_agent_for_command
 from imbue.mngr.cli.common_opts import add_common_options
 from imbue.mngr.cli.common_opts import setup_command_context
@@ -30,7 +31,7 @@ class CaptureCliOptions(CommonCliOptions):
     "--start/--no-start",
     default=True,
     show_default=True,
-    help="Automatically start the host/agent if stopped",
+    help="Automatically start the host and agent if offline/stopped",
 )
 @optgroup.option(
     "--full/--no-full",
@@ -51,12 +52,17 @@ def capture(ctx: click.Context, **kwargs: Any) -> None:
         mngr_ctx=mngr_ctx,
         address=opts.agent,
         host_filter=None,
-        is_start_desired=opts.start,
     )
     if result is None:
         return
 
-    agent, _host = result
+    host_ref, agent_ref = result
+    agent, _host = ensure_host_and_agent_started(
+        host_ref=host_ref,
+        agent_ref=agent_ref,
+        allow_auto_start=opts.start,
+        mngr_ctx=mngr_ctx,
+    )
 
     logger.debug("Capturing pane content for agent: {}", agent.name)
     content = agent.capture_pane_content(include_scrollback=opts.full)
