+++ libs/mngr/imbue/mngr/cli/connect_test.py
@@ -1,10 +1,10 @@
 """Unit tests for the connect CLI command."""
 
+from imbue.mngr.cli.agent_selector import build_status_text
+from imbue.mngr.cli.agent_selector import filter_agents
+from imbue.mngr.cli.agent_selector import handle_search_key
 from imbue.mngr.cli.connect import ConnectCliOptions
 from imbue.mngr.cli.connect import _build_connection_options
-from imbue.mngr.cli.connect import build_status_text
-from imbue.mngr.cli.connect import filter_agents
-from imbue.mngr.cli.connect import handle_search_key
 from imbue.mngr.config.data_types import MngrContext
 from imbue.mngr.primitives import AgentAddress
 from imbue.mngr.primitives import AgentLifecycleState
