+++ libs/mngr/imbue/mngr/cli/provision.py
@@ -9,6 +9,7 @@ from loguru import logger
 from imbue.mngr.api.provision import provision_agent
 from imbue.mngr.cli.address_params import AGENT_ADDRESS
 from imbue.mngr.cli.address_params import HOST_ADDRESS
+from imbue.mngr.cli.agent_utils import ensure_host_started_and_resolve_agent
 from imbue.mngr.cli.agent_utils import find_agent_for_command
 from imbue.mngr.cli.common_opts import add_common_options
 from imbue.mngr.cli.common_opts import setup_command_context
@@ -37,6 +38,7 @@ class ProvisionCliOptions(CommonCliOptions):
     # Behavior options
     bootstrap: str | None
     destroy_on_fail: bool
+    start: bool
     restart: bool
     # Provisioning options
     extra_provision_command: tuple[str, ...]
@@ -87,6 +89,12 @@ def _output_result(agent_name: str, output_opts: OutputOptions) -> None:
     "destroy_on_fail",
     default=False,
     help="Destroy the host if provisioning fails [future]",
+)
+@optgroup.option(
+    "--start/--no-start",
+    default=True,
+    show_default=True,
+    help="Automatically start the host if offline (the agent does not need to be running)",
 )
 @optgroup.option(
     "--restart/--no-restart",
@@ -147,19 +155,24 @@ def provision(ctx: click.Context, **kwargs: Any) -> None:
         raise UserInputError("Cannot specify both positional agent and --agent option")
     address: AgentAddress | None = opts.agent if opts.agent is not None else opts.agent_option
 
-    # Find the agent (start the host if needed, but don't require the agent to be running)
+    # Find the agent (provision needs the host online but manages the agent's
+    # state itself via --restart, so the agent process state is not required).
     result = find_agent_for_command(
         mngr_ctx=mngr_ctx,
         address=address,
         host_filter=None,
-        is_start_desired=True,
-        skip_agent_state_check=True,
     )
     if result is None:
         logger.info("No agent selected")
         return
 
-    agent, host = result
+    host_ref, agent_ref = result
+    agent, host = ensure_host_started_and_resolve_agent(
+        host_ref=host_ref,
+        agent_ref=agent_ref,
+        allow_auto_start=opts.start,
+        mngr_ctx=mngr_ctx,
+    )
 
     # Parse provisioning options
     provisioning = AgentProvisioningOptions(
@@ -194,7 +207,7 @@ def provision(ctx: click.Context, **kwargs: Any) -> None:
 CommandHelpMetadata(
     key="provision",
     one_line_description="Re-run provisioning on an existing agent [experimental]",
-    synopsis="mngr [provision|prov] [AGENT] [--agent <AGENT>] [--extra-provision-command <CMD>] [--upload-file <LOCAL:REMOTE>] [--env <KEY=VALUE>] [--env-file <FILE>] [--[no-]restart]",
+    synopsis="mngr [provision|prov] [AGENT] [--agent <AGENT>] [--extra-provision-command <CMD>] [--upload-file <LOCAL:REMOTE>] [--env <KEY=VALUE>] [--env-file <FILE>] [--start/--no-start] [--[no-]restart]",
     description="""This re-runs the provisioning steps (plugin lifecycle hooks, file transfers,
 custom commands, env vars) on an agent that has already been created. Useful for
 syncing configuration, authentication, and installing additional packages. Most
