+++ libs/mngr_kanpan/imbue/mngr_kanpan/fetcher.py
@@ -16,6 +16,7 @@ from imbue.imbue_common.pure import pure
 from imbue.mngr.api.discover import discover_hosts_and_agents
 from imbue.mngr.api.find import find_one_agent
 from imbue.mngr.api.list import list_agents
+from imbue.mngr.cli.agent_utils import ensure_host_started_and_resolve_agent
 from imbue.mngr.config.data_types import MngrContext
 from imbue.mngr.interfaces.data_types import AgentDetails
 from imbue.mngr.primitives import AgentAddress
@@ -223,10 +224,12 @@ def compute_section(fields: dict[str, FieldValue]) -> BoardSection:
 
 def toggle_agent_mute(mngr_ctx: MngrContext, agent_name: AgentName) -> bool:
     """Toggle the mute state of an agent. Returns the new mute state."""
-    agent, _host = find_one_agent(
-        AgentAddress(agent=agent_name),
-        mngr_ctx,
-        skip_agent_state_check=True,
+    host_ref, agent_ref = find_one_agent(AgentAddress(agent=agent_name), mngr_ctx)
+    agent, _host = ensure_host_started_and_resolve_agent(
+        host_ref=host_ref,
+        agent_ref=agent_ref,
+        allow_auto_start=False,
+        mngr_ctx=mngr_ctx,
     )
     plugin_data = agent.get_plugin_data(PLUGIN_NAME)
     is_muted = not plugin_data.get("muted", False)
